Moving unit tests to correct place
