+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/offline/repository/domain/channel/internal/ChannelRepositoryImplTest.kt
@@ -1,0 +1,64 @@
+@file:OptIn(ExperimentalCoroutinesApi::class)
+
+package io.getstream.chat.android.offline.repository.domain.channel.internal
+
+import io.getstream.chat.android.offline.randomChannel
+import io.getstream.chat.android.offline.randomMessage
+import io.getstream.chat.android.offline.randomUser
+import io.getstream.chat.android.test.TestCoroutineRule
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.runTest
+import org.junit.Rule
+import org.junit.jupiter.api.Test
+import org.mockito.kotlin.argThat
+import org.mockito.kotlin.doReturn
+import org.mockito.kotlin.mock
+import org.mockito.kotlin.verify
+import org.mockito.kotlin.whenever
+import java.util.Date
+
+internal class ChannelRepositoryImplTest {
+    @get:Rule
+    val testCoroutines: TestCoroutineRule = TestCoroutineRule()
+
+    private val channelDao: ChannelDao = mock()
+    private val channelRepository: ChannelRepositoryImpl =
+        ChannelRepositoryImpl(
+            channelDao,
+            { randomUser() },
+            { randomMessage() }
+        )
+
+    @Test
+    fun `Given channel without messages in DB Should insert channel with updated last message`() = runTest {
+        val channel = randomChannel(messages = emptyList())
+        val lastMessage = randomMessage(createdAt = Date())
+        whenever(channelDao.select(listOf("cid"))) doReturn listOf(channel.toEntity())
+
+        channelRepository.updateLastMessageForChannel("cid", lastMessage)
+
+        verify(channelDao).insert(argThat { channelEntity ->
+            channelEntity.lastMessageAt == lastMessage.createdAt &&
+                channelEntity.lastMessageId == lastMessage.id
+        })
+    }
+
+    @Test
+    fun `Given channel with outdated lastMessage in DB Should insert channel with updated last message`() = runTest {
+        val before = Date(1000)
+        val after = Date(2000)
+        val outdatedMessage = randomMessage(id = "messageId1", createdAt = before)
+        val newLastMessage = randomMessage(id = "messageId2", createdAt = after)
+        val channel = randomChannel(messages = listOf(outdatedMessage), lastMessageAt = before)
+        whenever(channelDao.select(listOf("cid"))) doReturn listOf(channel.toEntity())
+
+        channelRepository.updateLastMessageForChannel("cid", newLastMessage)
+
+        // channel.lastMessageAt `should be equal to` newLastMessage.createdAt
+        // channel.lastMessage `should be equal to` newLastMessage
+        verify(channelDao).insert(argThat { channelEntity ->
+            channelEntity.lastMessageAt == after &&
+                channelEntity.lastMessageId == newLastMessage.id
+        })
+    }
+}
